Curd bảng genre

+++ WebComic/Repositories/GenreRepository.cs
@@ -21,22 +21,35 @@ namespace WebComic.Repositories
 
         public async Task<Genre> GetGenreById(int id)
         {
-            return await _context.Genres.FindAsync(id);
+            Genre result = await _context.Genres.FirstOrDefaultAsync(g => g.GenreId == id);
+            if(result == null)
+            {
+                throw new InvalidOperationException("Thể loại không tồn tại");
+            }
+            return result;
         }
 
         public async Task UpdateGenre(int id, Genre genre)
         {
             if (id != genre.GenreId)
             {
-                throw new ArgumentException("ID mismatch");
+                throw new InvalidOperationException("ID không trngf khớp");
             }
-
+            Genre g = _context.Genres.FirstOrDefault(g => g.Name == genre.Name && g.GenreId != genre.GenreId);
+            if (g != null)
+            {
+                throw new InvalidOperationException("Thể loại đã tồn tại");
+            }
             _context.Entry(genre).State = EntityState.Modified;
             await _context.SaveChangesAsync();
         }
 
         public async Task<Genre> CreateGenre(Genre genre)
         {
+            if(genre.Name == null || genre.Name == "")
+            {
+                throw new InvalidOperationException("Không được để trống!");
+            }
             // Kiểm tra xem tên thể loại đã tồn tại chưa
             var existingGenre = await _context.Genres.FirstOrDefaultAsync(g => g.Name == genre.Name);
 
@@ -57,8 +70,18 @@ namespace WebComic.Repositories
             var genre = await _context.Genres.FindAsync(id);
             if (genre != null)
             {
+                List<ComicGenre> comicgenre =  _context.ComicGenres.Where(i => i.GenreId == id).ToList();
+                if (comicgenre.Count() > 0)
+                {
+                    throw new InvalidOperationException("Không thể xóa thể loại này");
+                }
                 _context.Genres.Remove(genre);
                 await _context.SaveChangesAsync();
+            }
+            else
+            {
+
+            throw new InvalidOperationException("Thể loại dã được xóa hoặc không tồn tại!");
             }
         }
     }
